requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0
streamlit>=1.29.0
python-dotenv>=1.0.0
pypdf>=3.17.4
tiktoken>=0.5.2
faiss-cpu>=1.7.4
numpy>=1.24.3
pandas>=2.0.3
